+ fix for @ComponentScan in Web Test Context + refactoring packages in web_test module
