+++ web/src/test/java/edu/softserve/zoo/web/test/controller/rest/EmployeeRestControllerTest.java
@@ -1,13 +1,13 @@
-package edu.softserve.zoo.controller;
+package edu.softserve.zoo.web.test.controller.rest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import edu.softserve.zoo.config.WebConfig;
 import edu.softserve.zoo.controller.rest.Routes;
 import edu.softserve.zoo.dto.TaskStatisticsDto;
 import edu.softserve.zoo.dto.TaskStatusDto;
 import edu.softserve.zoo.dto.TaskTypeDto;
 import edu.softserve.zoo.model.Task;
 import edu.softserve.zoo.util.AppProfiles;
+import edu.softserve.zoo.web.test.config.WebTestConfig;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -31,7 +31,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
  * @author Taras Zubrei
  */
 @RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = WebConfig.class)
+@ContextConfiguration(classes = WebTestConfig.class)
 @ActiveProfiles(AppProfiles.TEST)
 @WebAppConfiguration
 public class EmployeeRestControllerTest {
